+++ src/types/delete-my-commands.types.ts
@@ -1,0 +1,11 @@
+import { BotCommandScope } from './bot-command-scope.types';
+
+export interface DeleteMyCommandsData {
+  scope?: BotCommandScope;
+  language_code?: string;
+}
+
+export interface DeleteMyCommandsResponse {
+  ok: boolean;
+  result: boolean;
+}