+++ src/types/general.types.ts
@@ -1,0 +1,6 @@
+export type ChatId = number | string;
+
+export interface BotCommand {
+  command: string;
+  description: string;
+}