+++ src/types/index.ts
@@ -1,7 +1,17 @@
+export * from './general.types';
+export * from './bot-command-scope.types';
+
 export * from './get-me.types';
+
 export * from './get-updates.types';
+
 export * from './set-webhook.types';
 export * from './delete-webhook.types';
 export * from './get-webhook-info.types';
+
 export * from './send-message.types';
 export * from './send-document.types';
+
+export * from './set-my-commands.types';
+export * from './delete-my-commands.types';
+export * from './get-my-commands.types';