+++ src/types/send-message.types.ts
@@ -1,7 +1,8 @@
+import { ChatId } from './general.types';
 import { UpdateMessage } from './get-updates.types';
 
 export interface SendMessageData {
-  chat_id: number;
+  chat_id: ChatId;
   text: string;
 }