+++ CHANGELOG.md
@@ -1,3 +1,17 @@
 # Changelog
 
 _This changelog follows the [keep a changelog](https://keepachangelog.com)_ format to maintain a human readable changelog.
+
+### [1.1.0](https://github.com/nikita-nuykin/tg-bot-api/compare/v1.0.4...v1.1.0) (2022-05-17)
+
+#### Added
+
+- setMyCommands method
+- deleteMyCommands method
+- getMyCommands method
+
+### [1.0.4](https://github.com/nikita-nuykin/tg-bot-api/compare/v1.0.0...v1.0.4) (2022-05-17)
+
+#### Added
+
+- sendDocument method